Move itemsPerPage function to a utils module
