+++ src/Components/ClustersListTable/ClustersListTable.spec.ct.js
@@ -28,6 +28,7 @@ import {
   checkPaginationValues,
   changePagination,
 } from '../../../cypress/utils/table';
+import { itemsPerPage } from '../../../cypress/utils/pagination';
 
 const data = props['data'];
 // add property name to clusters
@@ -73,20 +74,6 @@ const TABLE_HEADERS_SORT_KEYS = {
   Low: 'low',
   'Last seen': 'last_seen',
 };
-
-// FIXME improve syntax
-// FIXME move to utils module
-function itemsPerPage() {
-  let items = data.length;
-  const array = [];
-  while (items > 0) {
-    const remain = items - DEFAULT_ROW_COUNT;
-    let v = remain > 0 ? DEFAULT_ROW_COUNT : items;
-    array.push(v);
-    items = remain;
-  }
-  return array;
-}
 
 // TODO: test pre-filled search parameters filtration
 
@@ -231,7 +218,7 @@ describe('clusters list table', () => {
       });
     });
     it('can iterate over pages', () => {
-      cy.wrap(itemsPerPage()).each((el, index, list) => {
+      cy.wrap(itemsPerPage(data.length)).each((el, index, list) => {
         checkRowCounts(TABLE, el).then(() => {
           // TODO why is this nested?
           expect(window.location.search).to.contain(
